Permissions update! ❌✔

+++ commands/remove.js
@@ -2,6 +2,9 @@
 const quotesSchema = require('../schemas/quotes-schema');
 const counterSchema = require('../schemas/counter-schema');
 const leaderboardSchema = require('../schemas/quote-leaderboard-schema');
+const { Permissions } = require('discord.js');
+
+// TODO: clean this up this mess
 
 module.exports = {
     category: 'MosesDB',
@@ -18,31 +21,69 @@ module.exports = {
     }],
 
 
-    callback: async({ interaction, args, user }) => {
+    callback: async({ interaction, args, user, member }) => {
 
         let response = '';
 
+
+        const quoteToEdit = await quotesSchema.findOne({
+            quoteId: args[0]
+        });
         // Check if document exists
         if (await quotesSchema.find({ quoteId: args }) != '') {
-            response = `Removed quote **\`${args}\`**.`;
 
-            // Deleting specified document from the "quotes" collection
-            await quotesSchema.deleteOne({
-                quoteId: args
-            });
+            if (!member.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) {
 
-            // Subtracting 1 from the auto increment counter in the "quote-counter" collection so that when you add a new quote it doesn't skip a number
-            // Like id does in fucking SQL
-            await counterSchema.updateOne({}, {
-                $inc: { seq_value: -1 }
-            });
+                if (quoteToEdit['submitterId'] == member.id) {
 
-            // Update the leaderboards cause no free points
-            await leaderboardSchema.updateOne({
-                userId: user.id
-            }, {
-                $inc: { count: -1 }
-            });
+                    // Deleting specified document from the "quotes" collection
+                    await quotesSchema.deleteOne({
+                        quoteId: args
+                    });
+
+                    // Subtracting 1 from the auto increment counter in the "quote-counter" collection so that when you add a new quote it doesn't skip a number
+                    // Like id does in SQL
+                    await counterSchema.updateOne({}, {
+                        $inc: { seq_value: -1 }
+                    });
+
+                    // Update the leaderboards cause no free points
+                    await leaderboardSchema.updateOne({
+                        userId: user.id
+                    }, {
+                        $inc: { count: -1 }
+                    });
+
+                    response += `Quote **#${quoteToEdit['quoteId']} →** "\`${quoteToEdit['quote']}\`" has been **permanently deleted**.`;
+
+                } else {
+
+                    response += `Unfortunately quote **#${args[0]}** was submitted by <@${quoteToEdit['submitterId']}>.\nIf you want it deleted ask them or a server admin.`;
+
+                }
+
+            } else {
+
+                await quotesSchema.deleteOne({
+                    quoteId: args
+                });
+
+                // Subtracting 1 from the auto increment counter in the "quote-counter" collection so that when you add a new quote it doesn't skip a number
+                // Like id does in SQL
+                await counterSchema.updateOne({}, {
+                    $inc: { seq_value: -1 }
+                });
+
+                // Update the leaderboards cause no free points
+                await leaderboardSchema.updateOne({
+                    userId: user.id
+                }, {
+                    $inc: { count: -1 }
+                });
+                response += `Quote **#${quoteToEdit['quoteId']} →** "\`${quoteToEdit['quote']}\`" has been **permanently deleted** (admin mode).`;
+
+            }
+
         } else {
             response = `Quote with the id **\`${args}\`** doesn't exist.`;
         }
